+++ ml_tools/ETL_cleaning.py
@@ -2,15 +2,17 @@ from ._core._ETL_cleaning import (
     save_unique_values,
     basic_clean,
     basic_clean_drop,
+    drop_macro_polars,
     DragonColumnCleaner,
     DragonDataFrameCleaner,
     info
 )
 
 __all__ = [
+    "DragonColumnCleaner",
+    "DragonDataFrameCleaner",
     "save_unique_values",
     "basic_clean",
     "basic_clean_drop",
-    "DragonColumnCleaner",
-    "DragonDataFrameCleaner"
+    "drop_macro_polars",
 ]